pyproject.toml
[tool.poetry]
name = "pytest-diff-selector"
version = "0.1.1"
description = "Get tests affected by code changes (using git)"
authors = ["Israel Fruchter <israel.fruchter@gmail.com>"]
license = "MIT"
readme = "README.md"

repository = "https://github.com/fruch/pytest-diff-selector"

keywords = ["pytest", "git"]

classifiers = [
    "Topic :: Software Development"
]

packages = [
    { include = "pytest_diff_selector" },
]

[tool.poetry.scripts]
selector = 'pytest_diff_selector.main:run'

[tool.poetry.dependencies]
python = "^3.6"
unidiff = "^0.7.3"
pyan3 = "^1.2.0"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-git = "^1.7.0"

[tool.poetry.plugins.pytest11]
pytest-diff-selector = "pytest_diff_selector.plugin"

[build-system]
requires = ["poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"
